scsi: dc395x: remove unncessary type casts
drivers/scsi/dc395x.c: In function '__dc395x_eh_bus_reset':
drivers/scsi/dc395x.c:97:36: error: value computed is not used [-Werror=unused-value]
   97 | #define DC395x_read8(acb,address)  (u8)(inb(acb->io_port_base + (address)))
      |                                    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
drivers/scsi/dc395x.c:1003:2: note: in expansion of macro 'DC395x_read8'
 1003 |  DC395x_read8(acb, TRM_S1040_SCSI_INTSTATUS);
      |  ^~~~~~~~~~~~
drivers/scsi/dc395x.c: In function 'data_io_transfer':
drivers/scsi/dc395x.c:97:36: error: value computed is not used [-Werror=unused-value]
   97 | #define DC395x_read8(acb,address)  (u8)(inb(acb->io_port_base + (address)))
      |                                    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ drivers/scsi/dc395x.c
@@ -94,9 +94,9 @@
 #define DC395x_LOCK_IO(dev,flags)		spin_lock_irqsave(((struct Scsi_Host *)dev)->host_lock, flags)
 #define DC395x_UNLOCK_IO(dev,flags)		spin_unlock_irqrestore(((struct Scsi_Host *)dev)->host_lock, flags)
 
-#define DC395x_read8(acb,address)		(u8)(inb(acb->io_port_base + (address)))
-#define DC395x_read16(acb,address)		(u16)(inw(acb->io_port_base + (address)))
-#define DC395x_read32(acb,address)		(u32)(inl(acb->io_port_base + (address)))
+#define DC395x_read8(acb,address)		inb(acb->io_port_base + (address))
+#define DC395x_read16(acb,address)		inw(acb->io_port_base + (address))
+#define DC395x_read32(acb,address)		inl(acb->io_port_base + (address))
 #define DC395x_write8(acb,address,value)	outb((value), acb->io_port_base + (address))
 #define DC395x_write16(acb,address,value)	outw((value), acb->io_port_base + (address))
 #define DC395x_write32(acb,address,value)	outl((value), acb->io_port_base + (address))
